requirements.txt
langchain
langserve
fastapi
uvicorn
openai
pandas
faiss-cpu
python-dotenv